feat(LPC_USART): Handle TXIDLE interrupt
Store the TXIDLE interrupt status and check if we need to generate an
interrupt in UpdateInterrupts().

We generate an interrupt if both conditions are true:
- The TXIDLE interrupt is enabled
- AND the TX queue is empty

This is part 1/2 of enabling the non-blocking API in the standard LPC
SDK fsl_usart.c driver.
Next we need to implement the TXIDLE status flag.

Signed-off-by: Laurens Miers <laurens.miers@mind.be>